pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "btcopilot"
version = "1.0.0"
description = "BT Copilot"
license = { file = "LICENSE" }
authors = [
    { name="Patrick Stinson", email="patrick@alaskafamilysystems.com" }
]


[project.optional-dependencies]
app = [
    "bcrypt",
    "celery==5.5.3",
    "chromadb",
    "click",
    "ddtrace",
    "gunicorn",
    "flask",
    "flask-bcrypt",
    "flask-cli",
    "flask-graphql",
    "flask-mail",
    "flask-sqlalchemy",
    "flask-wtf",
    # "graphene-sqlalchemy",
    "openai",
    "psycopg2-binary",
    "numpy<2",
    "vedana @ git+https://github.com/patrickkidd/vedana.git@release",

    # pro

    "pygments",
    "python-dateutil",
    "stripe",
    "alembic",
    "cachetools",
    "rich",
    "sip==6.8.6",
    "pyqt5==5.15.11",
    "pyqt5-sip==12.15.0",
    "pyqt5-qt5==5.15.2",
    "sentence_transformers",
    # "torch @ https://download.pytorch.org/whl/cpu/torch-2.4.1%2Bcpu-cp311-cp311-manylinux1_x86_64.whl",  # CPU-only PyTorch
    "spacy",
    "mistralai",
    "pymupdf4llm",

    # personal
    "pydantic_ai",

    # training
    "assemblyai",
    "chromadb",
    "langchain",
    "langchain-chroma",
    "langchain-community",
    "langchain-core",
    "langchain-huggingface",
    "langchain-openai",
    "langchain-text-splitters",
    "nest_asyncio",
    "pydantic_ai",
    "pypdf",
]

test = [
    "black",
    "build",
    "freezegun",
    "mock",
    "playwright",
    "pytest",
    "pytest-asyncio",
    "pytest-playwright",
    "requests",
    "snapshottest",
]


[project.urls]
repository = "https://github.com/patrickkidd/btcopilot"

[tool.setuptools]
packages = ["btcopilot"]
